present: allow line-wrapping of bullet list items
This is not strictly necessary, but being able to wrap these long lines
helps make the diffs for the Markdown conversion of old files easier to read.

The wrapping of the blog is in CL 222839.

For golang/go#33955.

Change-Id: I26c3f8db6b137c194f03b2538f221aa4fc3f2324
Reviewed-on: https://go-review.googlesource.com/c/tools/+/222843
Run-TryBot: Russ Cox <rsc@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Rob Pike <r@golang.org>