go/packages: fix compatibility with gopackagesdriver
The Mode field is part of an RPC interface with gopackagesdriver
implementations, so renumbering the Need bits by adding a new value in
the middle is a compatibility break. Move it to the bottom.

Also update an incorrect comment.

Change-Id: Idad9e9e8b163f3b516ff9055d503af8749dd343c
Reviewed-on: https://go-review.googlesource.com/c/tools/+/234099
Run-TryBot: Heschi Kreinick <heschi@google.com>
Reviewed-by: Rebecca Stambler <rstambler@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>