[dev.go2go] all: support the go2-dev go/types library in x/tools
Most notably, this change allows gopls to compile with the new go/types,
allowing it to work with the generics prototype.

This change will fail on TryBots, as it requires a new version of the
Go standard library. This will only be used to tag a go2go-compatible
version of gopls.

Change-Id: I6b43272d70dc4e92e151df259b5003b353c97d30
Reviewed-on: https://go-review.googlesource.com/c/tools/+/238258
Reviewed-by: Heschi Kreinick <heschi@google.com>