go/analysis/analysistest: Ignore error when...
...changed file cannot be formatted

There are instances when the diagnostics are reported correctly by the
analyzer, but the suggested fixes are wrong, and result in code that
cannot be formatted by `format.Source` (maybe due to syntax errors in
the edited code).  Right now, in those cases, the test runner ignores
the failed tests and moves on to the next test, giving an impression
that the tests have passed. This commit fixes that.

Signed-off-by: Sourya Vatsyayan <sourya@deepsource.io>