present: properly render section heading using the section's level number
This fixes https://github.com/golang/go/issues/6846

If we take the example from the readme:

Some Text

- bullets
- more bullets
- a bullet continued
  on the next line

Some More text

	Preformatted text (code block)
	is indented (by one tab, or four spaces)

The sub-section and sub-subsection will have the headings <h2> and <h3> respectively,
which breaks the page, since the main title has a heading of <h3> so it pushes down the sub-section to
the bottom of the page.

If we use .Level for the section heading, they will have the headings <h3> and <h4> which will render properly.