go/analysis/analysistest: fix issue in RunWithSuggestedFixes on a package with tests
The issue is described further in golang/go#40574.  The fix is simply to do the
suggested-fixes checking on each result separately.  We still get
duplicate messages, which is arguably the correct behavior but certainly
not so bad, and we no longer get a bunch of spurious errors.

Fixes golang/go#40574.