present: new css class for the heading title for slide articles
The .title class is used in case the title is the only thing being
rendered in an article. The position: absolute property was previously
set on the h2 element, which broke the article's contents.

Now the slide's articles will begin from h2 instead of h3, and will be properly
positioned.