always send diagnostics-done message for check cmd
When running the check cmd from the cli, it may hang indefinitely on a channel waiting for "diagnostics-done". The nonstandardRequest implementation currently does not always send "diagnostics-done" if there are errors. This fixes that, however it also throws out any error from diagnostics-done itself (which would unlikely to be useful).