cmd/present: embed static and template files
Prior to this commit, the present tool relied on static and template
files in ${GOPATH}/src/golang.org/x/tools/cmd/present.

As of Go 1.16 that directory is not populated when the present tool is
installed using the go get command.

This commit embeds those assets in the present tool binary using
embed.FS.

To acheive this, we moved both static and template data to a "base"
directory in the source.  This way, embedding "base" will include both
templates and static assets.  We wrap the resulting embedded filesystem
with fs.Sub() to remove the "base" prefix.