gopls/internal/hooks: create included licenses text
We're adding a command that shows the licenses of all the code we depend
on. Start by creating a (very kludgy) shell script to gather them into a
string constant. Hopefully something nicer will come along later, but
for now this works fine with our current dependencies.

There's no strong reason this needs to be in the gopls module but it
felt right and it needed to live somewhere.

Trybots don't have network access, so testing that the file is in sync
is a little tricky. For simplicity's sake, we bake the gopls go.sum into
the generated file, and if that goes out of sync we assume the licenses
need updating. That's an overapproximation, since a dependency update
doesn't necessarily mean a license change, but it's easy and we don't
update deps often.

Change-Id: I338834044078f7df25304fa50864716057ce308e
Reviewed-on: https://go-review.googlesource.com/c/tools/+/281858
Trust: Heschi Kreinick <heschi@google.com>
Run-TryBot: Heschi Kreinick <heschi@google.com>
gopls-CI: kokoro <noreply+kokoro@google.com>
TryBot-Result: Go Bot <gobot@golang.org>
Reviewed-by: Rebecca Stambler <rstambler@golang.org>
Reviewed-by: Russ Cox <rsc@golang.org>