flag to show internal pkg, respect go.mod module
additional flag per command line to show the internal
packages per default and not only by '?m=all'. This is
only enabled for the 'Standard library' but not 'Third party'

identify the go.mod module if running in GOMOD mode. If
detected check the RootType based on module path in
moduleFS. Last exclude the /cmd/ from modules package
view.