x/tools/go/analysis/passes/copylock: detect copylock in multi-assignment
The existing implementation does not identify lock copies when doing
multi-assignment, i.e. foo, ok := someMap["key"]

Currently, it will find issue with similar code foo := someMap["key"].

This discrepency is because a return type of Tuple was not considered
when resolving the type of the right hand side of the assignment.

Fixes #45896