cmd/bundle: quote command-line arguments
The previous version of bundle simply joins together with a space " ".
While this works for simple arguments this causes problems when you
need to pass special strings like an empty string or a string containing
a space.

The following example shows how the previous version handles an empty
string argument `-prefix ""`.

    //go:generate bundle -o /dev/stdout -prefix  golang.org/x/tools/cmd/bundle

This change quotes the arguments with strconv.Quote, if needed, before joining
together with a space:

    //go:generate bundle -o /dev/stdout -prefix "" golang.org/x/tools/cmd/bundle