godoc/static: Update subrepo tour links
The tour links under Sub-repositories on the main page of the godoc server
are pointing at the old module (outside of x/website) and the old (pre
go.dev) public URL. Should be updated to match the new homes of the tour.

Update packageroot.html to point at the new tour package doc in
x/website/tour and generate the godoc static content.

Fixes golang/go#59505