gopls/internal/lsp/source: fix rename spurious shadowing error
This change fixes a bug in the rename lexical shadowing check
due to an edge case in the identifier resolution mechanism.
In func(x T) {}, the scope of the function is the entire thing,
and the reference to T appears within it, but logically
it is resolved in the enclosing environment.

(This has always been a subtle an error-prone aspect
of the types.Scope representation.)

Also, add a test for the four cases (receiver, type param,
param, result).

Fixes golang/go#60752

Change-Id: Ic047977df37696f6d70c375a227e1e91fa29f681
Reviewed-on: https://go-review.googlesource.com/c/tools/+/509560
Run-TryBot: Alan Donovan <adonovan@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>
gopls-CI: kokoro <noreply+kokoro@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>