gopls/internal/lsp/cache: make options immutable on the View
In the zero-config gopls world, it will be much simpler if Views are
immutable. As a first step toward this, make options an immutable
property of the View, and store them in a new Folder abstraction. In the
future, folders may be shared across multiple views.

This removes the 'minorOptionsChange' logic that avoided recreating
Views if the options change was not 'major'. Digging into the history of
that functionality, it seems like a premature optimization, or perhaps
it was only necessary for tests. Notably, the old completion marker
tests relied on being able to toggle options in between assertions (or
in some cases, in the middle of assertions!). Those tests ran too slow
on this CL, and so have all been ported over to the new marker
framework, which uses static per-suite options. On average, most real
clients should change options very infrequently if at all, and so
reinitializing the view (effectively restarting gopls) is not a big
deal.

For golang/go#57979

Change-Id: I78d86a9ddb85f29e781705b898a66a5642ff2611
Reviewed-on: https://go-review.googlesource.com/c/tools/+/538796
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>