go/ssa: TestStdlib: report count of syntactic functions too
The existing AllFunctions measure includes the result of
(what turns out to be) a rather complex and underspecified
transitive visitation. This change adds a SrcFunctions
metric that corresponds to FuncDecl and FuncLit syntax,
which is important for the performance of
go/analysis/passes/buildssa.

Also, add TestNetHTTP which exercises SSA building of a
package without its dependencies, which corresponds to
the package-at-a-time workloads seen in go/analysis.

Change-Id: If32ea01b7f9d6a0ee886b062ed9218e867a103c2
Reviewed-on: https://go-review.googlesource.com/c/tools/+/538435
Auto-Submit: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Tim King <taking@google.com>